requirements.txt
flask==2.0.1
tensorflow==2.15.0
opencv-python-headless==4.9.0.80
numpy==1.24.4
werkzeug==2.0.1
ffmpeg-python==0.2.0
keras==2.15.0
transformers==4.40.0  # For NLP model
torch==2.3.0          # For transformers
scikit-learn==1.4.0   # For text processing
nltk==3.8.1           # For text preprocessing